fixed minesweeper module for new block families
