+++ module.txt
@@ -1,6 +1,6 @@
 {
     "id" : "MineSweeper",
-    "version" : "0.1.2-SNAPSHOT",
+    "version" : "0.1.3-SNAPSHOT",
     "author" : "",
     "displayName" : "MineSweeper",
     "description" : "",
